requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.20
djangorestframework==3.14.0
exceptiongroup==1.1.2
idna==3.4
importlib-metadata==6.7.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
pyshorteners==1.0.1
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
siege==0.0.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
zipp==3.15.0
